:truck: move serializer fields to a separate file
